% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBioCyc.R
\name{cyc2Tax}
\alias{cyc2Tax}
\title{Get NCBI Taxonomy ID From BioCyc ID}
\usage{
cyc2Tax(cycID, n = 4)
}
\arguments{
\item{cycID}{The cyc species ID. The KEGG support multiple species ID, for example c('HUMAN', 'MOUSE', 'ECOLI').}

\item{n}{The number of CPUs or processors, and the default value is 4.}
}
\value{
The corresponding NCBI Taxonomy ID in character vector.
}
\description{
BioCyc Database API - Get the NCBI taxonomy ID from a given BioCyc ID
}
\details{
NCBI taxonomy ID is used as unique ID accoss cyc and BioCyc databases. This functions is used to get the corresponding NCBI Taxonomy ID from BioCyc.If BioCyc has no official NCBI taxonomy ID, it will return a character with length of 0 (it looks like "").
}
\examples{
# get human, mouse, and Ecoli NCBI taxonomy ID with 2 threads
cyc2Tax(c('HUMAN', 'MOUSE', 'ECOLI'), n = 2)
# transfer all BioCyc species ID to NCBI taxonomy ID
\dontrun{
wBiocycSpe <- getPhyloCyc(whole = TRUE)
wNCBISpe <- cyc2Tax(wBiocycSpe[, 1])
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

