% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGGAdd.R
\name{getKEGGMotifList}
\alias{getKEGGMotifList}
\title{Get motif list from KEGG.}
\usage{
getKEGGMotifList(motifName)
}
\arguments{
\item{motifName}{A single KEGG motif ID}
}
\value{
A matrix of KEGG genes and description
}
\description{
Get genes processing certain motif
}
\details{
Get all the gene having certain motif.
}
\examples{
\dontrun{
getKEGGMotifList('pf:DUF3675')}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

