% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGG.R
\name{getTIDSeqFasta}
\alias{getTIDMulSeqFasta}
\alias{getTIDSeqFasta}
\title{Get protein and gene senqeces from T numbers}
\usage{
getTIDSeqFasta(TID, seqType = "aaseq")

getTIDMulSeqFasta(TIDs, n = 4, ...)
}
\arguments{
\item{TID}{The T number ID for the protein or gene.}

\item{seqType}{Choose nucleotide acid ('ntseq') or amino acid ('aaseq') seqences, and the default is amino acid sequences.}

\item{TIDs}{A vector of T number IDs.}

\item{n}{The number of CPUs or processors, and the default value is 4.}

\item{...}{Parameters inherited from getTIDSeqFasta()}
}
\value{
A BStringSet

A BStringSet
}
\description{
Get nucleotide acid and amino acid sequences according to the T numbers
}
\details{
Get protein and gene sequences from KEGG T number in fasta format. As there is no direct API for retrieving the sequence from T number, for example "T10017:100009". The fasta sequence is extract from a webpage like "http://www.genome.jp/dbget-bin/www_bget?-f+-n+a+t10017:100009". The function getTIDSeqFasta() get a sequence one time, and the function getTIDMulSeqFasta() provides a parallel way to download sequences.
}
\examples{
# get the nucleotide sequence of "T10017:100009"
getTIDSeqFasta('T10017:100009', seqType = 'ntseq')
getTIDMulSeqFasta(c('T10017:100009', 'T10017:100036', 'T10017:100044'), n = 2)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}

